Fix issues with class-label webhook
The current issue with the assign-class-label webhook is that if the
webhook pod goes down and the deployment tries to spin it back up, it
fails. This is because the MutatingWebhookConfiguration is trying to
call the webhook on the webhook pod, but there is no webhook pod so it
fails to create. This commit
makes it so that the MutatingWebhookConfiguration does not call the
webhook on the webhook pod, that way if the pod goes down and tries to
spin back up it is able to without failing because there is no webhook
pod.

This commit also changes replicas for the webhook pod to 2.
Also adds the secret.yaml to the kustomization.yaml file.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>